create singletons containers
