+++ src/test/java/com/rbkmoney/magista/config/AbstractDaoConfig.java
@@ -8,32 +8,16 @@ import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.TestPropertySource;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
-
-import java.util.stream.Stream;
-
-import static io.github.benas.randombeans.EnhancedRandomBuilder.aNewEnhancedRandom;
+import org.springframework.transaction.annotation.Transactional;
 
 @SpringBootTest
-@Testcontainers
 @ContextConfiguration(
         classes = MagistaApplication.class,
         initializers = AbstractDaoConfig.Initializer.class)
 @DirtiesContext
 @TestPropertySource("classpath:application.yml")
-public abstract class AbstractDaoConfig {
-
-    private static final String POSTGRESQL_IMAGE_NAME = "postgres";
-    private static final String POSTGRESQL_VERSION = "11.4";
-
-    @Container
-    public static final PostgreSQLContainer<?> POSTGRESQL_CONTAINER = new PostgreSQLContainer<>(
-            DockerImageName
-                    .parse(POSTGRESQL_IMAGE_NAME)
-                    .withTag(POSTGRESQL_VERSION));
+@Transactional
+public abstract class AbstractDaoConfig extends AbstractDbSingletonContainerConfig {
 
     public static class Initializer extends ConfigDataApplicationContextInitializer {
 
@@ -55,13 +39,5 @@ public abstract class AbstractDaoConfig {
                     "payouter.pooling.enabled=false")
                     .applyTo(configurableApplicationContext);
         }
-    }
-
-    public static <T> T random(Class<T> type, String... excludedFields) {
-        return aNewEnhancedRandom().nextObject(type, excludedFields);
-    }
-
-    public static <T> Stream<T> randomStreamOf(int amount, Class<T> type, String... excludedFields) {
-        return aNewEnhancedRandom().objects(type, amount, excludedFields);
     }
 }
