+++ src/test/java/com/rbkmoney/magista/dao/AdjustmentDaoTest.java
@@ -8,6 +8,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.List;
 
+import static com.rbkmoney.magista.util.RandomBeans.random;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class AdjustmentDaoTest extends AbstractDaoConfig {
