+++ src/test/java/com/rbkmoney/magista/dao/ChargebackDaoTest.java
@@ -9,6 +9,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import java.util.Collections;
 import java.util.List;
 
+import static com.rbkmoney.magista.util.RandomBeans.random;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class ChargebackDaoTest extends AbstractDaoConfig {
