+++ src/test/java/com/rbkmoney/magista/dao/InvoiceTemplateDaoTest.java
@@ -7,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 
 import java.util.Collections;
 
+import static com.rbkmoney.magista.util.RandomBeans.random;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class InvoiceTemplateDaoTest extends AbstractDaoConfig {
