+++ src/test/java/com/rbkmoney/magista/dao/PaymentDaoTest.java
@@ -13,6 +13,8 @@ import java.util.UUID;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import static com.rbkmoney.magista.util.RandomBeans.random;
+import static com.rbkmoney.magista.util.RandomBeans.randomStreamOf;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class PaymentDaoTest extends AbstractDaoConfig {
