+++ src/test/java/com/rbkmoney/magista/service/AdjustmentServiceTest.java
@@ -14,6 +14,8 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static com.rbkmoney.magista.util.RandomBeans.random;
+import static com.rbkmoney.magista.util.RandomBeans.randomStreamOf;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class AdjustmentServiceTest extends AbstractDaoConfig {
