+++ src/test/java/com/rbkmoney/magista/service/ChargebackServiceTest.java
@@ -15,6 +15,8 @@ import org.springframework.boot.test.mock.mockito.SpyBean;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static com.rbkmoney.magista.util.RandomBeans.random;
+import static com.rbkmoney.magista.util.RandomBeans.randomStreamOf;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
