+++ src/test/java/com/rbkmoney/magista/service/RefundServiceTest.java
@@ -12,6 +12,8 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import java.util.List;
 import java.util.stream.Collectors;
 
+import static com.rbkmoney.magista.util.RandomBeans.random;
+import static com.rbkmoney.magista.util.RandomBeans.randomStreamOf;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.BDDMockito.given;
